Fix relative path invocation and remove default sample.xml
Previously, when starting docbookrx from the source root as "bin/docbookrx"
it would throw an exception:
/usr/share/rubygems/rubygems/core_ext/kernel_require.rb:54:in `require':
 cannot load such file -- bin/../lib/docbookrx (LoadError)

Also removed the attempt to read sample.xml as default.  The "Please specify a
DocBook file to convert" message would never appear; the script would complain
it could not find sample.xml instead.

And finally: will complain that non-xml files are not valid instead of raising a nil
exception.