Fix buggy output from some admonition conversions
The assumption that single a single "para" inside the note
guarantees that "inline" NOTE can be used is not correct.
Instead of trying to fix the test, I changed the code to
always use the safe format instead.  Try to convert this example
with the old code to see the bug:

<note>
  <para>
  Please note the fruit:
  <screen>Apple, oranges and bananas</screen>
  </para>
</note>